+++ lua/obsidian/commands/new.lua
@@ -1,5 +1,8 @@
 ---@param data obsidian.CommandArgs
 return function(data)
   local id = data.args:len() > 0 and data.args
-  require("obsidian.actions").new(id)
+  ---@diagnostic disable-next-line: param-type-mismatch
+  require("obsidian.actions").new(id, function(note)
+    note:open { sync = true }
+  end)
 end